refactor(sideboard_guide_panel): split SideboardGuidePanel into mixin package
Move widgets/panels/sideboard_guide_panel.py into a package: frame.py
builds the UI; properties.py holds the translate helper, getters, and
the pure-data card-list formatter; handlers.py owns the entry setters,
refresh, and event callbacks.
