+++ widgets/panels/sideboard_guide_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Sideboard guide panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.sideboard_guide_panel.frame import SideboardGuidePanel
+
+__all__ = ["SideboardGuidePanel"]
